Move main Duke class to duke package
