+++ build.gradle
@@ -29,7 +29,7 @@ test {
 }
 
 application {
-    mainClassName = "Duke"
+    mainClassName = "duke.Duke"
 }
 
 shadowJar {
