+++ src/main/java/duke/data/Storage.java
@@ -1,4 +1,4 @@
-package data;
+package duke.data;
 
 import java.io.IOException;
 import java.nio.charset.Charset;
