+++ src/main/java/duke/data/TaskList.java
@@ -1,6 +1,6 @@
-package data;
+package duke.data;
 
-import task.Task;
+import duke.task.Task;
 
 import java.io.IOException;
 import java.util.ArrayList;
