+++ src/main/java/duke/task/Deadline.java
@@ -1,4 +1,4 @@
-package task;
+package duke.task;
 
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
