+++ src/test/java/duke/task/EventTest.java
@@ -1,4 +1,4 @@
-package task;
+package duke.task;
 
 import org.junit.jupiter.api.Test;
